Records - Allow parent entity field
Users can now add the Parent entity field directly into the record's structure, this field is typically visible at the top of the editor labelled "Child record of" but the value couldn't be edited. If added to the entity structure, the field will now appear where placed instead of at the top.

Upon attempting to save the child record with changes to the parent entity field, a check is performed to see if the parent record has a field that is using the child record pointer. If it is used, the parent record value will be restored with a message explaining why.