Field entry mask - Initial version
Added the ability to define an field entry mask for record fields.
This will allow users to define a simple mask applied to the field's value that follow a general pattern, e.g. for links to external records.
The mask can be added via the record editor for freetext and numeric fields